requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Data & APIs
yfinance>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0

# Trading
alpaca-py>=0.9.0

# Utilities
click>=8.1.0
pyyaml>=6.0
